+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/DependencyManager.groovy
@@ -1,6 +1,7 @@
 package life.qbic.portal.offermanager
 
 import groovy.util.logging.Log4j2
+import life.qbic.business.offers.fetch.FetchOffer
 import life.qbic.datamodel.dtos.business.AcademicTitle
 import life.qbic.datamodel.dtos.business.AffiliationCategory
 import life.qbic.business.persons.affiliation.create.CreateAffiliation
@@ -9,6 +10,8 @@ import life.qbic.business.offers.create.CreateOffer
 import life.qbic.datamodel.dtos.business.Offer
 import life.qbic.datamodel.dtos.general.Person
 import life.qbic.portal.offermanager.communication.EventEmitter
+import life.qbic.portal.offermanager.components.offer.overview.OfferOverviewController
+import life.qbic.portal.offermanager.components.offer.overview.OfferOverviewPresenter
 import life.qbic.portal.offermanager.components.person.search.SearchPersonView
 import life.qbic.portal.offermanager.components.person.search.SearchPersonViewModel
 import life.qbic.portal.offermanager.components.person.update.UpdatePersonViewModel
@@ -82,6 +85,7 @@ class DependencyManager {
     private CreateAffiliationPresenter createAffiliationPresenter
     private CreateOfferPresenter createOfferPresenter
     private CreateOfferPresenter updateOfferPresenter
+    private OfferOverviewPresenter offerOverviewPresenter
 
     private PersonDbConnector customerDbConnector
     private OfferDbConnector offerDbConnector
@@ -93,6 +97,9 @@ class DependencyManager {
     private CreateAffiliation createAffiliation
     private CreateOffer createOffer
     private CreateOffer updateOffer
+    private FetchOffer fetchOfferOfferOverview
+    private FetchOffer fetchOfferCreateOffer
+    private FetchOffer fetchOfferUpdateOffer
 
     private CreatePersonController createCustomerController
     private CreatePersonController updateCustomerController
@@ -100,6 +107,7 @@ class DependencyManager {
     private CreateAffiliationController createAffiliationController
     private CreateOfferController createOfferController
     private CreateOfferController updateOfferController
+    private OfferOverviewController offerOverviewController
 
     private CreatePersonView createCustomerView
     private CreatePersonView updatePersonView
@@ -323,15 +331,26 @@ class DependencyManager {
         } catch (Exception e) {
             log.error("Unexpected exception during ${CreateOfferViewModel.getSimpleName()} setup", e)
         }
+        try {
+            this.offerOverviewPresenter = new OfferOverviewPresenter(this.viewModel, this.offerOverviewModel)
+        } catch (Exception e) {
+            log.error("Unexpected exception during ${OfferOverviewPresenter.getSimpleName()} setup", e)
+        }
     }
 
     private void setupUseCaseInteractors() {
         this.createCustomer = new CreatePerson(createCustomerPresenter, customerDbConnector)
         this.createCustomerNewOffer = new CreatePerson(createCustomerPresenterNewOffer, customerDbConnector)
+
         this.createAffiliation = new CreateAffiliation(createAffiliationPresenter, customerDbConnector)
+
         this.createOffer = new CreateOffer(offerDbConnector, createOfferPresenter)
         this.updateOffer = new CreateOffer(offerDbConnector, updateOfferPresenter)
         this.updateCustomer = new CreatePerson(updateCustomerPresenter, customerDbConnector)
+
+        this.fetchOfferOfferOverview = new FetchOffer(offerDbConnector, offerOverviewPresenter)
+        this.fetchOfferCreateOffer = new FetchOffer(offerDbConnector, createOfferPresenter)
+        this.fetchOfferUpdateOffer = new FetchOffer(offerDbConnector, updateOfferPresenter)
     }
 
     private void setupControllers() {
@@ -360,14 +379,19 @@ class DependencyManager {
             throw e
         }
         try {
-            this.createOfferController = new CreateOfferController(this.createOffer,this.createOffer)
+            this.createOfferController = new CreateOfferController(this.createOffer, this.fetchOfferCreateOffer, this.createOffer)
         } catch (Exception e) {
             log.error("Unexpected exception during ${CreateOfferController.getSimpleName()} setup", e)
         }
         try {
-            this.updateOfferController = new CreateOfferController(this.updateOffer,this.updateOffer)
+            this.updateOfferController = new CreateOfferController(this.updateOffer, this.fetchOfferUpdateOffer, this.updateOffer)
         } catch (Exception e) {
             log.error("Unexpected exception during ${CreateOfferController.getSimpleName()} setup", e)
+        }
+        try {
+            this.offerOverviewController = new OfferOverviewController(this.fetchOfferOfferOverview)
+        } catch (Exception e) {
+            log.error("Unexpected exception during ${OfferOverviewController.getSimpleName()} setup", e)
         }
     }
 
@@ -431,7 +455,7 @@ class DependencyManager {
 
         OfferOverviewView overviewView
         try {
-            overviewView = new OfferOverviewView(offerOverviewModel)
+            overviewView = new OfferOverviewView(offerOverviewModel, offerOverviewController)
         } catch (Exception e) {
             log.error("Could not create ${OfferOverviewView.getSimpleName()} view.", e)
             throw e